packages/python-sdk/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentdoor"
version = "0.1.0"
description = "Python Agent SDK for AgentDoor - headless agent authentication"
requires-python = ">=3.9"
license = "MIT"
dependencies = [
    "httpx>=0.25.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "respx>=0.20", "pytest-cov>=4.0"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.coverage.run]
source = ["agentdoor"]

[tool.coverage.report]
fail_under = 80
show_missing = true
